Split EAR platform-broker suite into its own spec
Rename ear-flows.spec.ts to ear-basic.spec.ts (web-flow EAR suite) and
move the platform-broker suite into ear-platform-broker.spec.ts as a
plain describe block. The platform spec is excluded from CI via the
pipeline testFilter (ear-basic) instead of describe.skip.

Copilot-Session: 7155fa61-5730-4d8d-9553-a89e8768e478
